+++ src/screens/ChatScreen.tsx
@@ -30,7 +30,7 @@ import { launchCamera, launchImageLibrary } from 'react-native-image-picker'
 import MaterialIcons from 'react-native-vector-icons/MaterialIcons'
 
 // TODO: setup absolute paths and eslint rules for import arrangement
-import { firebaseFunctions } from '../helpers/firebaseFunctions'
+import { firebaseFunctions } from '../utils/firebaseFunctions'
 
 import type { RootStackParamList } from '../../App'
 import type { Message, ChatRoom } from '../types/server'
